open_chat/requirements.txt
dj-database-url==0.5.0
Django==3.0.1
djangorestframework==3.11.0
requests==2.7.0
gunicorn==19.7.1
Markdown==2.6.9
Pillow==2.3.0
django-filter==1.0.4
django-heroku==0.3.1